extracting catch_up logic
