+++ node/src/reactor/main_reactor/reactor_state.rs
@@ -1,6 +1,7 @@
 use datasize::DataSize;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use std::fmt::{Display, Formatter};
 
 #[derive(Copy, Clone, PartialEq, Eq, DataSize, Debug, Deserialize, JsonSchema, Serialize)]
 pub enum ReactorState {
@@ -16,4 +17,17 @@ pub enum ReactorState {
     Validate,
     // node should be shut down for upgrade
     ShutdownForUpgrade,
+}
+
+impl Display for ReactorState {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ReactorState::Initialize => write!(f, "Initialize"),
+            ReactorState::CatchUp => write!(f, "CatchUp"),
+            ReactorState::Upgrading => write!(f, "Upgrading"),
+            ReactorState::KeepUp => write!(f, "KeepUp"),
+            ReactorState::Validate => write!(f, "Validate"),
+            ReactorState::ShutdownForUpgrade => write!(f, "ShutdownForUpgrade"),
+        }
+    }
 }
